+++ _Typings/gameApi/types/collider.d.ts
@@ -18,8 +18,8 @@ declare module "game:type" {
 
     type CollisionEvent = {
         eventType: "Enter" | "Exit" | "Stay"
-        itemA: Item
-        itemB: Item
+        itemA: Item | Player
+        itemB: Item | Player
         impulse: float
         contactPoint: Float3
         normal: Float3