+++ CHANGELOG.md
@@ -11,8 +11,9 @@ All notable changes to this project will be documented in this file.
 -   Restricted the value range of `levelManager.cameraDirection` to `0 | 1 | 2 | 3`.
 -   Restricted the value range of the parameter `which` of `uiElement.onPointerClick()` to `0 | 1 | 2`.
 -   let `AssertGuid` accept `string` as a valid value when the type parameter is not `const`.
--   Now the `getTransform()` returns `[Float3, Float3, Float3]` instead of `Float3[]`.
+-   Now `getTransform()` returns `[Float3, Float3, Float3]` instead of `Float3[]`.
 -   Added a new type `Tuple` in `utils.d.ts` and applied it where needed.
+-   Now the type of `triggeredItem` of `OnTrigger` is `Item | Player` instead of `Item`.
 
 ## v1.0.6 (2025-05-29)